requirements.txt
absl-py
academictorrents
addict
aiofiles
aiohttp
aiosignal
albumentations
altair
antlr4-python3-runtime
anyio
async-timeout
attrs
# backports.zoneinfo
basicsr
bencode.py
bitstring
blinker
blis
braceexpand
brotlipy
cachetools
catalogue
certifi
cffi
charset-normalizer
click
coloredlogs
confection
contourpy
cryptography
cycler
cymem
decorator
einops
emoji
entrypoints
fastapi
ffmpy
filelock
fitter
flatbuffers
flit-core
fonttools
frozenlist
fsspec
ftfy
future
fvcore
gitdb
GitPython
google-auth
google-auth-oauthlib
gradio
graphviz
grpcio
h11
httpcore
httpx
huggingface-hub
humanfriendly
idna
imageio
imageio-ffmpeg
importlib-metadata
importlib-resources
imutils
invisible-watermark
iopath
Jinja2
joblib
jsonschema
kiwisolver
kornia
langcodes
linkify-it-py
lmdb
lpips
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdit-py-plugins
mdurl
mkl-fft
mkl-random
mkl-service
mpmath
multidict
murmurhash
networkx
nltk
numpy
oauthlib
omegaconf
onnx
onnxruntime
open-clip-torch
opencv-contrib-python
opencv-python
opencv-python-headless
orjson
packaging
pandas
pathy
Pillow
pkgutil-resolve-name
portalocker
preshed
prettytable
protobuf
pyarrow
pyasn1
pyasn1-modules
pycparser
pycryptodome
pydantic
pydeck
pyDeprecate
pydub
Pygments
Pympler
pyOpenSSL
pyparsing
PyPubSub
pyrsistent
PySocks
python-dateutil
python-multipart
pytorch-lightning
pytz
pytz-deprecation-shim
PyWavelets
PyYAML
qudida
regex
requests
requests-oauthlib
rfc3986
rich
rsa
safetensors
SceneGraphParser
scikit-image
scikit-learn
scipy
semver
six
smart-open
smmap
sniffio
spacy
spacy-legacy
spacy-loggers
srsly
stanza
starlette
streamlit
streamlit-drawable-canvas
sympy
tabulate
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
termcolor
test-tube
thinc
threadpoolctl
tifffile
timm
tokenizers
toml
toolz
torch
torchmetrics
torchvision
torchviz
tornado
tqdm
transformers
typer
typing-extensions
tzlocal
uc-micro-py
urllib3
uvicorn
validators
wasabi
watchdog
wcwidth
webdataset
websockets
Werkzeug
yacs
yapf
yarl
zipp
# absl-py==1.4.0
# academictorrents==2.3.3
# addict==2.4.0
# aiofiles==23.1.0
# aiohttp==3.8.4
# aiosignal==1.3.1
# albumentations==1.3.0
# altair==4.2.2
# antlr4-python3-runtime==4.8
# anyio==3.6.2
# async-timeout==4.0.2
# attrs==22.2.0
# backports.zoneinfo==0.2.1
# basicsr==1.4.2
# bencode.py==2.0.0
# bitstring==3.1.5
# blinker==1.5
# blis==0.7.9
# braceexpand==0.1.7
# brotlipy==0.7.0
# cachetools==5.3.0
# catalogue==2.0.8
# certifi @ file:///croot/certifi_1671487769961/work/certifi
# cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
# charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
# click==8.1.3
# coloredlogs==15.0.1
# confection==0.0.4
# contourpy==1.0.7
# cryptography @ file:///croot/cryptography_1673298753778/work
# cycler==0.11.0
# cymem==2.0.7
# decorator==5.1.1
# einops==0.3.0
# emoji==2.2.0
# entrypoints==0.4
# fastapi==0.92.0
# ffmpy==0.3.0
# filelock==3.9.0
# fitter==1.5.2
# flatbuffers==23.1.21
# flit-core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
# fonttools==4.38.0
# frozenlist==1.3.3
# fsspec==2023.1.0
# ftfy==6.1.1
# future==0.18.3
# fvcore==0.1.5.post20221221
# gitdb==4.0.10
# GitPython==3.1.31
# google-auth==2.16.0
# google-auth-oauthlib==0.4.6
# gradio==3.16.2
# graphviz==0.20.1
# grpcio==1.51.1
# h11==0.14.0
# httpcore==0.16.3
# httpx==0.23.3
# huggingface-hub==0.12.0
# humanfriendly==10.0
# idna @ file:///croot/idna_1666125576474/work
# imageio==2.9.0
# imageio-ffmpeg==0.4.2
# importlib-metadata==6.0.0
# importlib-resources==5.10.2
# imutils==0.5.4
# invisible-watermark==0.1.5
# iopath==0.1.10
# Jinja2==3.1.2
# joblib==1.2.0
# jsonschema==4.17.3
# kiwisolver==1.4.4
# kornia==0.6.0
# langcodes==3.3.0
# linkify-it-py==1.0.3
# lmdb==1.4.0
# lpips==0.1.4
# Markdown==3.4.1
# markdown-it-py==2.1.0
# MarkupSafe==2.1.2
# matplotlib==3.7.0
# mdit-py-plugins==0.3.3
# mdurl==0.1.2
# mkl-fft==1.3.1
# mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
# mkl-service==2.4.0
# mpmath==1.2.1
# multidict==6.0.4
# murmurhash==1.0.9
# networkx==3.0
# nltk==3.7
# numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
# oauthlib==3.2.2
# omegaconf==2.1.1
# onnx==1.13.0
# onnxruntime==1.14.0
# open-clip-torch==2.0.2
# opencv-contrib-python==4.3.0.36
# opencv-python==4.7.0.68
# opencv-python-headless==4.7.0.68
# orjson==3.8.6
# packaging==23.0
# pandas==1.5.3
# pathy==0.10.1
# Pillow==9.3.0
# pkgutil-resolve-name==1.3.10
# portalocker==2.7.0
# preshed==3.0.8
# prettytable==3.6.0
# protobuf==3.20.3
# pyarrow==11.0.0
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
# pycryptodome==3.17
# pydantic==1.10.5
# pydeck==0.8.0
# pyDeprecate==0.3.1
# pydub==0.25.1
# Pygments==2.14.0
# Pympler==1.0.1
# pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
# pyparsing==3.0.9
# PyPubSub==3.3.0
# pyrsistent==0.19.3
# PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
# python-dateutil==2.8.2
# python-multipart==0.0.5
# pytorch-lightning==1.5.0
# pytz==2022.7.1
# pytz-deprecation-shim==0.1.0.post0
# PyWavelets==1.4.1
# PyYAML==6.0
# qudida==0.0.4
# regex==2022.10.31
# requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
# requests-oauthlib==1.3.1
# rfc3986==1.5.0
# rich==13.3.1
# rsa==4.9
# safetensors==0.2.7
# SceneGraphParser==0.1.0
# scikit-image==0.19.3
# scikit-learn==1.2.1
# scipy==1.10.0
# # Editable install with no version control (segment-anything==1.0)
# -e /home/cz2021/project/segment-anything-main
# semver==2.13.0
# six @ file:///tmp/build/80754af9/six_1644875935023/work
# smart-open==6.3.0
# smmap==5.0.0
# sniffio==1.3.0
# spacy==3.5.2
# spacy-legacy==3.0.12
# spacy-loggers==1.0.4
# srsly==2.4.6
# stanza==1.4.2
# starlette==0.25.0
# streamlit==1.12.1
# streamlit-drawable-canvas==0.8.0
# sympy==1.11.1
# tabulate==0.9.0
# tensorboard==2.12.0
# tensorboard-data-server==0.7.0
# tensorboard-plugin-wit==1.8.1
# termcolor==2.2.0
# test-tube==0.7.5
# thinc==8.1.9
# threadpoolctl==3.1.0
# tifffile==2023.2.3
# timm==0.6.12
# tokenizers==0.12.1
# toml==0.10.2
# toolz==0.12.0
# torch==1.12.1
# torchmetrics==0.6.0
# torchvision==0.13.1
# torchviz==0.0.2
# tornado==6.2
# tqdm==4.64.1
# transformers==4.19.2
# typer==0.7.0
# typing-extensions @ file:///croot/typing_extensions_1669924550328/work
# tzdata==2022.7
# tzlocal==4.2
# uc-micro-py==1.0.1
# urllib3 @ file:///croot/urllib3_1673575502006/work
# uvicorn==0.20.0
# validators==0.20.0
# wasabi==1.1.1
# watchdog==2.2.1
# wcwidth==0.2.6
# webdataset==0.2.5
# websockets==10.4
# Werkzeug==2.2.3
# yacs==0.1.8
# yapf==0.32.0
# yarl==1.8.2
# zipp==3.13.0